git-svn-id: https://svn.qdecoder.org/qdecoder/trunk@269 5a34604f-89d6-db11-b6c0-0015c5f490bc

+++ src/qObstack.c
@@ -137,7 +137,7 @@ Q_OBSTACK *qObstackInit(void) {
  *
  * @since not released yet
  */
-bool qObstackGrow(Q_OBSTACK *obstack, const void *object, int size) {
+bool qObstackGrow(Q_OBSTACK *obstack, const void *object, size_t size) {
 	if(obstack == NULL || object == NULL || size <= 0) return false;
 	return qEntryPut(obstack->stack, "", object, size, false);
 }
